Fixed a problem with skipping sources.
Now the skip position is set to the maximum possible value. Relying of
the source length is not secure; it may be incorrect.

+++ playerEngine/src/main/scala/de/oliver_heger/linedj/player/engine/impl/ArraySourceImpl.scala
@@ -47,10 +47,10 @@ private object ArraySourceImpl {
    * @param offsetDelta the start offset of the data to be selected
    * @return the adjusted ''ArraySource''
    */
-  def apply(source: ArraySource, offsetDelta: Int): ArraySource =
+  def apply(source: ArraySource, offsetDelta: Long): ArraySource =
     if(offsetDelta <= 0) source
-    else new ArraySourceImpl(source.data, offset = source.offset + offsetDelta,
-      length = math.max(source.length - offsetDelta, 0))
+    else new ArraySourceImpl(source.data, offset = (source.offset + offsetDelta).toInt,
+      length = math.max(source.length - offsetDelta, 0).toInt)
 
   /**
     * Creates a new ''ArraySource'' from a ''ByteString''. The resulting
